llm/api/requirements.txt
asyncio
websockets
fastapi
openai==1.7.2
python-dotenv==1.0.0